added git action yml
