+++ .github/workflows/dotnet-tests.yml
@@ -1,0 +1,28 @@
+name: .NET Tests
+
+on:
+  push:
+    branches: [ main, labs ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Setup .NET
+        uses: actions/setup-dotnet@v4
+        with:
+          dotnet-version: 8.0.x
+
+      - name: Restore dependencies
+        run: dotnet restore ./library.sln
+
+      - name: Build
+        run: dotnet build ./library.sln --no-restore --configuration Release
+
+      - name: Test
+        run: dotnet test ./Tests/Tests.csproj --no-build --verbosity normal
