+++ library/library.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "library.AppHost", "library.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "library.ServiceDefaults", "library.ServiceDefaults\library.ServiceDefaults.csproj", "{9D0FB5CC-97E4-B8BE-2B2C-2062B059C91A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{503A006B-D3AE-45D0-934B-5533381A9DE9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{9D0FB5CC-97E4-B8BE-2B2C-2062B059C91A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9D0FB5CC-97E4-B8BE-2B2C-2062B059C91A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9D0FB5CC-97E4-B8BE-2B2C-2062B059C91A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{503A006B-D3AE-45D0-934B-5533381A9DE9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{503A006B-D3AE-45D0-934B-5533381A9DE9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{503A006B-D3AE-45D0-934B-5533381A9DE9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{503A006B-D3AE-45D0-934B-5533381A9DE9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
